fix: 缩放倍数设置类型限定符丢小数点位丢失导致的排版计算错误
添加小数点支持在dfmstyleditemdelegate中

Log:
Bug: https://pms.uniontech.com/zentao/bug-view-69706.html

Former-commit-id: 3aa57eb9a301b08e8e1f9d7eedf74b824be51c40

+++ src/dde-file-manager-lib/interfaces/dfmstyleditemdelegate.cpp
@@ -302,10 +302,11 @@ QPixmap DFMStyledItemDelegate::getIconPixmap(const QIcon &icon, const QSize &siz
         }
     }
 
+    //类型限定符的更改会导致缩放小数点丢失，从而引发缩放因子的bug
     if (px.width() > icon_size.width() * pixelRatio) {
-        px.setDevicePixelRatio(px.width() / icon_size.width());
+        px.setDevicePixelRatio(px.width() / qreal(icon_size.width()));
     } else if (px.height() > icon_size.height() * pixelRatio) {
-        px.setDevicePixelRatio(px.height() / icon_size.height());
+        px.setDevicePixelRatio(px.height() / qreal(icon_size.height()));
     } else {
         px.setDevicePixelRatio(pixelRatio);
     }
